requirements.txt
requests==2.32.3
pandas==2.2.2
pyarrow==17.0.0
python-dotenv==1.0.1
pydantic==2.8.2
pytest==8.3.2
